backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
pdfplumber>=0.10.0
python-docx>=1.1.0
geopy>=2.4.0
joblib>=1.3.0
python-dotenv>=1.0.0
tensorflow>=2.10.0
requests>=2.31.0

